chore: split sync runner fallback tests
